Response format specification on /summary
On POST requests, uses the format form field and on GET requests the
format query parameter.

In both cases, valid values are `html` and `json`. If an invalid value
is passed, returns a 400 response specifying what the appropriate values
are.

`html` is the default value if none is specified. This keeps things
consistent for apps using Thumbsup today (search frontend renders
Thumbsup in iframe for TL;DR feature).

Resolves https://github.com/simiotics/thumbsup/issues/20